chat - polish chat status after move
